requirements.txt
streamlit>=1.32.0
google-generativeai>=0.5.0
gspread>=6.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
cloudinary>=1.36.0
pandas>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
